Shuffle namespace tests
